// src/types/drugs.ts
export interface DrugsData {
  drug_tests_total: number;
  charges: number;
  arrests: number;
  year: number;
  start_date: string;
  end_date: string;
  jurisdiction: string;
  location: string;
  age_group: string;
  best_detection_method: string;
  detection_method: string;
  amphetamine: string;
  cannabis: string;
  cocaine: string;
  ecstasy: string;
  methylamphetamine: string;
  other: string;
  unknown: string;
  no_drugs_detected: string;
  positive_drug_tests: number;
  fines: number;
}